Updated the TextFieldRich
